Add files via upload
Specific Forecast hours range for PRDGEN per FAA product request has been added as following
1. IFI AK 3km: up to f018
2. IFI G130 13km: up to f021
3. GTG G130 13km: up to f021

ATM part
4. G130 13km: up to f021
5. G237 32km: up to f012
6. rotate 13km prslev: up to f015
7. rotate 13km natlev: up to f006

"subset_ifi_304m.sh" will 
 1.  extract the 3 variables in IFI file at every 304 m on forecast hour (1, 2, 3, 6, 9, 12, 15, 18)
 2. each variable has its own output filename
 3. all 3 files will be in a directory "wmo" under the RRFS workflow default output directory for PRDGEN.  WMO header will will added when available

+++ ush/prdgen/rrfs_prdgen_faa_subpiece.sh
@@ -12,6 +12,8 @@ ififip=$5
 aviati=$6
 COMOUT=$7
 
+USHrrfs=$USHdir/prdgen
+
 # FAA request variable to be extracted from UPP output
   parmdir=${FIXprdgen}
   # parmdir=${USHdir}/../fix/prdgen
@@ -21,7 +23,18 @@ COMOUT=$7
 
   #-- fcst string in wrgib2 inv file. Set to "" for all ins & ave
   fcstvar1="${ifhr} hour fcst"
-  fcstvar2="0-${ifhr} hour acc fcst"
+
+  #-- hourly average instead of accumulation from "0"
+  if [ ${ifhr} = 0 ] ; then
+    jfhr=0
+  else    
+    let jfhr=${ifhr}-1
+  fi 
+
+  fcstvar2="${jfhr}-${ifhr} hour acc fcst"
+
+  #-- accumulation from "0"
+  # fcstvar2="0-${ifhr} hour acc fcst"
 
   #-- replace undifined variables in "*parmas" files in /fix/prdgen
   sed "s/FCSTVARS1/${fcstvar1}/" ${parmdir}/rrfs.prslev-FAA130.params > rrfs.prslev-FAA130.params
@@ -53,30 +66,36 @@ COMOUT=$7
   if [[ -f ${COMOUT}/${prslev} ]]; then
 
     #-- GRID 130: CONUS 13km
-    wgrib2 ${COMOUT}/${prslev} | grep -F -f rrfs.prslev-FAA130.params | \
-    wgrib2 -i ${COMOUT}/${prslev} -set_bitmap 1 -set_grib_type c3 \
-       -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-       -new_grid_interpolation bilinear \
-       -new_grid ${grid_specs_130} rrfs.t${cyc}z.prslev.f${fhr}.conus_13km.faa.grib2
-    mv rrfs.t${cyc}z.prslev.f${fhr}.conus_13km.faa.grib2 ${COMOUT}
-    wgrib2 ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.conus_13km.faa.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.conus_13km.faa.grib2.idx
+    if [ $ifhr -le 21 ]; then
+      wgrib2 ${COMOUT}/${prslev} | grep -F -f rrfs.prslev-FAA130.params | \
+      wgrib2 -i ${COMOUT}/${prslev} -set_bitmap 1 -set_grib_type c3 \
+         -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+         -new_grid_interpolation bilinear \
+         -new_grid ${grid_specs_130} rrfs.t${cyc}z.prslev.f${fhr}.conus_13km.faa.grib2
+      mv rrfs.t${cyc}z.prslev.f${fhr}.conus_13km.faa.grib2 ${COMOUT}
+      wgrib2 ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.conus_13km.faa.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.conus_13km.faa.grib2.idx
+    fi
 
     #-- 13km Rotated Lat Lon
-    wgrib2 ${COMOUT}/${prslev} | grep -F -f rrfs.prslev-rrfs13km.params | \
-    wgrib2 -i ${COMOUT}/${prslev} -set_bitmap 1 -set_grib_type c3 \
-       -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-       -new_grid_interpolation bilinear \
-       -new_grid ${grid_specs_rrfs_13km} rrfs.t${cyc}z.prslev.f${fhr}.rotate_13km.faa.grib2
-    mv rrfs.t${cyc}z.prslev.f${fhr}.rotate_13km.faa.grib2 ${COMOUT}
-    wgrib2 ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.rotate_13km.faa.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.rotate_13km.faa.grib2.idx
+    if [ $ifhr -le 15 ]; then
+      wgrib2 ${COMOUT}/${prslev} | grep -F -f rrfs.prslev-rrfs13km.params | \
+      wgrib2 -i ${COMOUT}/${prslev} -set_bitmap 1 -set_grib_type c3 \
+         -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+         -new_grid_interpolation bilinear \
+         -new_grid ${grid_specs_rrfs_13km} rrfs.t${cyc}z.prslev.f${fhr}.rotate_13km.faa.grib2
+      mv rrfs.t${cyc}z.prslev.f${fhr}.rotate_13km.faa.grib2 ${COMOUT}
+      wgrib2 ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.rotate_13km.faa.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.rotate_13km.faa.grib2.idx
+    fi
 
     #-- GRID 237: PR 32 km
-    wgrib2 ${COMOUT}/${prslev} | grep -F -f rrfs.prslev-FAA237.params | \
-    wgrib2 -i ${COMOUT}/${prslev} -set_bitmap 1 -set_grib_type c3 \
-       -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-       -new_grid_interpolation bilinear \
-       -new_grid ${grid_specs_237} rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2
-    mv rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2 ${COMOUT}
+    if [ $ifhr -le 12 ]; then
+      wgrib2 ${COMOUT}/${prslev} | grep -F -f rrfs.prslev-FAA237.params | \
+      wgrib2 -i ${COMOUT}/${prslev} -set_bitmap 1 -set_grib_type c3 \
+         -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+         -new_grid_interpolation bilinear \
+         -new_grid ${grid_specs_237} rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2
+      mv rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2 ${COMOUT}
+    fi
   fi
 
   if [[ -f ${COMOUT}/${natlev} ]]; then
@@ -90,40 +109,44 @@ COMOUT=$7
     #mv rrfs.t${cyc}z.natlev.f${fhr}.conus_13km.faa.grib2 ${COMOUT}
 
     #-- 13km Rotated Lat Lon
-    wgrib2 ${COMOUT}/${natlev} -s | grep "hybrid level:" | grep -F -f ${parmdir}/rrfs.natlev-FAA130.params | \
-    wgrib2 -i ${COMOUT}/${natlev} -set_bitmap 1 -set_grib_type c3 \
-       -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-       -new_grid_interpolation bilinear \
-       -new_grid ${grid_specs_rrfs_13km} rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2
+    if [ $ifhr -le 6 ]; then
+      wgrib2 ${COMOUT}/${natlev} -s | grep "hybrid level:" | grep -F -f ${parmdir}/rrfs.natlev-FAA130.params | \
+      wgrib2 -i ${COMOUT}/${natlev} -set_bitmap 1 -set_grib_type c3 \
+         -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+         -new_grid_interpolation bilinear \
+         -new_grid ${grid_specs_rrfs_13km} rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2
 
-    wgrib2 ${COMOUT}/${natlev} -s | grep ":LTNG:entire atmosphere:" | \
-    wgrib2 -i ${COMOUT}/${natlev} -set_bitmap 1 -set_grib_type c3 \
-       -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-       -new_grid_interpolation bilinear \
-       -new_grid ${grid_specs_rrfs_13km} rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.tmp.grib2
-    cat rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.tmp.grib2 >> rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2
-    rm rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.tmp.grib2
-    mv rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2 ${COMOUT}
-    wgrib2 ${COMOUT}/rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2.idx
+      wgrib2 ${COMOUT}/${natlev} -s | grep ":LTNG:entire atmosphere:" | \
+      wgrib2 -i ${COMOUT}/${natlev} -set_bitmap 1 -set_grib_type c3 \
+         -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+         -new_grid_interpolation bilinear \
+         -new_grid ${grid_specs_rrfs_13km} rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.tmp.grib2
+      cat rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.tmp.grib2 >> rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2
+      rm rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.tmp.grib2
+      mv rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2 ${COMOUT}
+      wgrib2 ${COMOUT}/rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.natlev.f${fhr}.rotate_13km.faa.grib2.idx
+    fi
 
     #-- GRID 237
-    #wgrib2 ${COMOUT}/${natlev} -s | grep "hybrid level:" | grep -F -f ${parmdir}/rrfs.natlev-FAA130.params | \
-    #wgrib2 -i ${COMOUT}/${natlev} -set_bitmap 1 -set_grib_type c3 \
-    #   -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-    #   -new_grid_interpolation bilinear \
-    #   -new_grid ${grid_specs_237} rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.faa.grib2
+    if [ $ifhr -le 12 ]; then
+      #wgrib2 ${COMOUT}/${natlev} -s | grep "hybrid level:" | grep -F -f ${parmdir}/rrfs.natlev-FAA130.params | \
+      #wgrib2 -i ${COMOUT}/${natlev} -set_bitmap 1 -set_grib_type c3 \
+      #   -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+      #   -new_grid_interpolation bilinear \
+      #   -new_grid ${grid_specs_237} rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.faa.grib2
 
-    wgrib2 ${COMOUT}/${natlev} -s | grep ":LTNG:entire atmosphere:" | \
-    wgrib2 -i ${COMOUT}/${natlev} -set_bitmap 1 -set_grib_type c3 \
-       -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-       -new_grid_interpolation bilinear \
-       -new_grid ${grid_specs_237} rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.tmp.grib2
-    #cat rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.tmp.grib2 >> rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.faa.grib2
-    cat rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.tmp.grib2 >> ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2
-    wgrib2 ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2.idx
+      wgrib2 ${COMOUT}/${natlev} -s | grep ":LTNG:entire atmosphere:" | \
+      wgrib2 -i ${COMOUT}/${natlev} -set_bitmap 1 -set_grib_type c3 \
+         -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+         -new_grid_interpolation bilinear \
+         -new_grid ${grid_specs_237} rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.tmp.grib2
+      #cat rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.tmp.grib2 >> rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.faa.grib2
+      cat rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.tmp.grib2 >> ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2
+      wgrib2 ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.pr_32km.faa.grib2.idx
 
-    rm rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.tmp.grib2
-    #mv rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.faa.grib2 ${COMOUT}
+      rm rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.tmp.grib2
+      #mv rrfs.t${cyc}z.natlev.f${fhr}.pr_32km.faa.grib2 ${COMOUT}
+    fi 
   fi
 
   #if [[ -f ${COMOUT}/rrfs.t${cyc}z.prslev.f${fhr}.conus_13km.faa.grib2  || -f ${COMOUT}/rrfs.t${cyc}z.natlev.f${fhr}.conus_13km.faa.grib2 ]]; then
@@ -150,14 +173,16 @@ COMOUT=$7
   # GTG process
 
   if [[ -f ${COMOUT}/${aviati} ]]; then
-    # wgrib2 ${COMOUT}/${aviati} -s | egrep '(:EDPARM:|:CATEDR:|:MWTURB:|:CITEDR:)' | \
-    wgrib2 ${COMOUT}/${aviati} -s | egrep '(:EDPARM:|:CATEDR:|:MWTURB:|:var discipline=0 master_table=2 parmcat=19 parm=50:)' | \
-    wgrib2 -i ${COMOUT}/${aviati} -set_bitmap 1 -set_grib_type c3 \
-       -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-       -new_grid_interpolation bilinear \
-       -new_grid ${grid_specs_130} GTG_grid_130.grib2
-    mv GTG_grid_130.grib2 ${COMOUT}/rrfs.t${cyc}z.aviati.f${fhr}.conus_13km.grib2
-    wgrib2 ${COMOUT}/rrfs.t${cyc}z.aviati.f${fhr}.conus_13km.grib2  -s > ${COMOUT}/rrfs.t${cyc}z.aviati.f${fhr}.conus_13km.grib2.idx
+    if [ $ifhr -le 21 ]; then
+      # wgrib2 ${COMOUT}/${aviati} -s | egrep '(:EDPARM:|:CATEDR:|:MWTURB:|:CITEDR:)' | \
+      wgrib2 ${COMOUT}/${aviati} -s | egrep '(:EDPARM:|:CATEDR:|:MWTURB:|:var discipline=0 master_table=2 parmcat=19 parm=50:)' | \
+      wgrib2 -i ${COMOUT}/${aviati} -set_bitmap 1 -set_grib_type c3 \
+         -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+         -new_grid_interpolation bilinear \
+         -new_grid ${grid_specs_130} GTG_grid_130.grib2
+      mv GTG_grid_130.grib2 ${COMOUT}/rrfs.t${cyc}z.aviati.f${fhr}.conus_13km.grib2
+      wgrib2 ${COMOUT}/rrfs.t${cyc}z.aviati.f${fhr}.conus_13km.grib2  -s > ${COMOUT}/rrfs.t${cyc}z.aviati.f${fhr}.conus_13km.grib2.idx
+    fi
   fi
 
   # IFI icing process
@@ -165,17 +190,27 @@ COMOUT=$7
   if [[ -f ${COMOUT}/${ififip} ]]; then
 
     #-- GRID 130: CONUS 13km
-    # wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:ICESEV:)' | \
-    wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:var discipline=0 master_table=2 parmcat=19 parm=37:)' | \
-    wgrib2 -i ${COMOUT}/${ififip} -set_bitmap 1 -set_grib_type c3 \
-       -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-       -new_grid_interpolation neighbor \
-       -new_grid ${grid_specs_130} IFI_grid_130.grib2
-    mv IFI_grid_130.grib2 ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.conus_13km.grib2
-    wgrib2 ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.conus_13km.grib2  -s > ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.conus_13km.grib2.idx
+    if [ $ifhr -le 21 ]; then
+      # wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:ICESEV:)' | \
+      wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:var discipline=0 master_table=2 parmcat=19 parm=37:)' | \
+      wgrib2 -i ${COMOUT}/${ififip} -set_bitmap 1 -set_grib_type c3 \
+         -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+         -new_grid_interpolation neighbor \
+         -new_grid ${grid_specs_130} IFI_grid_130.grib2
+      mv IFI_grid_130.grib2 ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.conus_13km.grib2
+      wgrib2 ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.conus_13km.grib2  -s > ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.conus_13km.grib2.idx
+
+      #-- subset IFI data at every 304 m only at certain forcast hour
+      if [ $ifhr = 1 -o  $ifhr = 2 -o  $ifhr = 3 -o  $ifhr = 6 -o  $ifhr = 9 -o  $ifhr = 12 -o  $ifhr = 15 -o  $ifhr = 18 ]; then
+        IFIFILE=rrfs.t${cyc}z.ififip.f${fhr}.conus_13km.grib2
+        IFIDOMAIN=conus_13km
+        ${USHrrfs}/subset_ifi_304m.sh $fhr $cyc ${COMOUT} ${IFIFILE} ${IFIDOMAIN}
+      fi
+    fi
 
     #-- 13km Rotated Lat Lon
-    #wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:var discipline=0 master_table=2 parmcat=19 parm=37:)' | \
+    #wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:ICESEV:)' | \
+    # wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:var discipline=0 master_table=2 parmcat=19 parm=37:)' | \
     #wgrib2 -i ${COMOUT}/${ififip} -set_bitmap 1 -set_grib_type c3 \
     #   -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
     #   -new_grid_interpolation neighbor \
@@ -184,20 +219,32 @@ COMOUT=$7
     #wgrib2 ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.rotate_13km.grib2  -s > ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.rotate_13km.grib2.idx
 
     #-- GRID 91: AK 3km
-    wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:var discipline=0 master_table=2 parmcat=19 parm=37:)' | \
-    wgrib2 ${COMOUT}/${ififip} -set_bitmap 1 -set_grib_type c3 \
-       -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
-       -new_grid_interpolation neighbor \
-       -new_grid ${grid_specs_91}  rrfs.t${cyc}z.ififip.f${fhr}.ak_3km.grib2
-    mv rrfs.t${cyc}z.ififip.f${fhr}.ak_3km.grib2 ${COMOUT}
-    wgrib2 ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.ak_3km.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.ak_3km.grib2.idx
+    if [ $ifhr -le 18 ]; then
+      # wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:ICESEV:)' | \
+      wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:var discipline=0 master_table=2 parmcat=19 parm=37:)' | \
+      wgrib2 ${COMOUT}/${ififip} -set_bitmap 1 -set_grib_type c3 \
+         -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
+         -new_grid_interpolation neighbor \
+         -new_grid ${grid_specs_91}  rrfs.t${cyc}z.ififip.f${fhr}.ak_3km.grib2
+      mv rrfs.t${cyc}z.ififip.f${fhr}.ak_3km.grib2 ${COMOUT}
+      wgrib2 ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.ak_3km.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.ak_3km.grib2.idx
+
+      #-- subset IFI data at every 304 m only at certain forcast hour
+      if [ $ifhr = 1 -o  $ifhr = 2 -o  $ifhr = 3 -o  $ifhr = 6 -o  $ifhr = 9 -o  $ifhr = 12 -o  $ifhr = 15 -o  $ifhr = 18 ]; then
+        IFIFILE=rrfs.t${cyc}z.ififip.f${fhr}.ak_3km.grib2
+        IFIDOMAIN=ak_3km
+        ${USHrrfs}/subset_ifi_304m.sh $fhr $cyc ${COMOUT} ${IFIFILE} ${IFIDOMAIN}
+      fi
+    fi
 
     #-- GRID 237: PR 32km
-    #wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:var discipline=0 master_table=2 parmcat=19 parm=37:)' | \
+    #wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:ICESEV:)' | \
+    # wgrib2 ${COMOUT}/${ififip} -s | egrep '(:ICPRB:|:SIPD:|:var discipline=0 master_table=2 parmcat=19 parm=37:)' | \
     #wgrib2 ${COMOUT}/${ififip} -set_bitmap 1 -set_grib_type c3 \
     #   -new_grid_winds grid -new_grid_vectors "UGRD:VGRD:USTM:VSTM" \
     #   -new_grid_interpolation neighbor \
     #   -new_grid ${grid_specs_237}  rrfs.t${cyc}z.ififip.f${fhr}.pr_32km.grib2
     #mv rrfs.t${cyc}z.ififip.f${fhr}.pr_32km.grib2 ${COMOUT}
     #wgrib2 ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.pr_32km.grib2 -s > ${COMOUT}/rrfs.t${cyc}z.ififip.f${fhr}.pr_32km.grib2.idx
+    
   fi
